requirements.txt
pyyaml>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-whois>=0.9.4
dnspython>=2.4.0
selenium>=4.20.0
selenium-stealth>=1.0.6
